move hashEmail to code-studio
