+++ code-studio/src/js/code-studio.js
@@ -36,6 +36,7 @@ var jquery = require('jquery-shim');
 window.dashboard = window.dashboard || {};
 window.dashboard.clientState = require('./clientState.js')(window.sessionStorage, jquery);
 window.dashboard.createCallouts = require('./callouts');
+window.dashboard.hashEmail = require('./hashEmail');
 
 // Wrap existing window onerror caller with a script error check.  If we have a
 // script error and a url, throw that so that we have the info in new relic.
