+++ dashboard/app/views/devise/registrations/new.html.haml
@@ -124,7 +124,7 @@
   $( "#user_school" ).placeholder();
 
   $( ".signupform" ).on("submit", function(e){
-    hashEmail({ email_selector: '#user_email',
-                hashed_email_selector: '#user_hashed_email',
-                age_selector: '#user_age'});
+    window.dashboard.hashEmail({ email_selector: '#user_email',
+      hashed_email_selector: '#user_hashed_email',
+      age_selector: '#user_age'});
   });
